+++ src/main/java/duke/Storage.java
@@ -1,3 +1,10 @@
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -10,7 +17,7 @@ public class Storage {
     private File file;
 
     /**
-     * Constructor for the Storage class.
+     * Constructor for the duke.Storage class.
      *
      * @param filePath The string representing the path of the file used.
      */
