+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,10 +1,18 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Todo;
+
 import java.io.IOException;
 
-public class AddTodoCommand extends Command{
+public class AddTodoCommand extends Command {
     private String taskName;
 
     /**
-     * Constructor for the AddTodoCommand class.
+     * Constructor for the duke.command.AddTodoCommand class.
      *
      * @param taskName
      */
