+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
