+++ src/main/java/duke/command/DeleteTaskCommand.java
@@ -1,24 +1,32 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
 import java.io.IOException;
 
-public class MarkNotDoneCommand extends Command {
+public class DeleteTaskCommand extends Command {
     private int index;
 
     /**
-     * Constructor for the MarkNotDoneCommand class.
+     * Constructor for the duke.command.DeleteTaskCommand class.
      *
-     * @param index Index of the task to be marked.
+     * @param index Index of the task to be deleted.
      */
-    public MarkNotDoneCommand(int index) {
+    public DeleteTaskCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         try {
-            taskList.markNotDone(index);
             Task task = taskList.getTask(index);
+            taskList.delete(index);
             storage.rewriteToFile(taskList.getList());
-            ui.successfulMarkNotDoneMsg(task.displayableForm());
+            ui.successfulTaskDeletionMsg(task.displayableForm(), taskList.getSize());
         } catch (IOException e) {
             throw new DukeException("Something went wrong: " + e.getMessage());
         }
